+++ recipes/hotfix-XS62ESP1027.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: xs_maintenance
-# Recipe:: hotfix-XS62ESP1025
+# Recipe:: hotfix-XS62ESP1027
 #
 # Copyright (C) 2015 Todd Pigram
 #
@@ -17,24 +17,24 @@
 # limitations under the License.
 #
 
-# VENOM - CVE-2015-3456: http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2015-3456
+# CVE-2015-4106, CVE-2015-4163, CVE-2015-4164, CVE-2015-2756, CVE-2015-4103, CVE-2015-4104, CVE-2015-4105
 
-bash "install XS62ESP1025" do
+bash "install XS62ESP1027" do
   user "root"
   cwd "/tmp"
   code <<-EOH
   mkdir -p /tmp/hotfixes
   cd /tmp/hotfixes
 
-  wget http://downloadns.citrix.com.edgesuite.net/10564/XS62ESP1025.zip
-  unzip XS62ESP1025.zip
+  wget http://downloadns.citrix.com.edgesuite.net/10661/XS62ESP1027.zip
+  unzip XS62ESP1027.zip
 
   . /etc/xensource-inventory
 
-  PATCHUUID=$(xe patch-upload file-name=XS62ESP1025.xsupdate)
+  PATCHUUID=$(xe patch-upload file-name=XS62ESP1027.xsupdate)
   xe patch-pool-apply uuid=${PATCHUUID}
 
   xe patch-clean uuid=${PATCHUUID}
   EOH
-  not_if {::File.exists?(node['xs62sp1']['025'])}
+  not_if {::File.exists?(node['xs62sp1']['027'])}
 end